requirements.txt
# Biology Tutorial Workflow - Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Model Providers
openai>=1.10.0
volcengine-python-sdk>=1.0.0  # 豆包/火山引擎

# Configuration
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async Support
aiofiles>=23.2.0
asyncio>=3.4.3

# Retry Mechanism
tenacity>=8.2.0

# Redis (Optional)
redis>=5.0.0
aioredis>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0

# Utilities
python-dotenv>=1.0.0
uuid>=1.30
